+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## v1.6.0
+- Add support for case-insensitive glob pattern matching ([#13](https://github.com/qetza/replacetokens/issues/13)).
+- Remove `readTextFile` function.
+
 ## v1.5.0
 - Externalize variable retrieval through callback to support more usage scenarios like external name normalization.
 - Rename `parseVariables` to `loadVariables` and add logs.